# Replace regex with faster rstrip