@aparrish 2015-06-24
looks good—in general I'm not too concerned about optimizing one-time calls like this, but the solution you propose works fine and isn't any more or less clear than the previous implementation.